* CalculatorExpr: now capability of String concatenation * CalculatorExpr: some docu and deprectated designation. * IteratorMask: also an array is Iterable, using {@link #IteratorMask(Object[], long)}. * OutTextPreparer: new in &lt;:for:var:container> {@link ForCmd} also an integer index   can be used as var if container is an Integer.   This is used for UFBgl code generation execution with array index. * util.TestOrg: new expect(double, double, acuracy) for numeric double and float results * XmlDataNode: only commented and with an additional assert, to clarify confusion.

+++ java/org/vishia/util/DataAccess.java
@@ -2448,10 +2448,10 @@ public class DataAccess {
     while(clazz1.isArray() && ixix < ixArray.length) {
       clazz1 = clazz1.getComponentType();
       if(clazz1 == Integer.TYPE){           //Note: the order is regarded to probability and calculation time. Integer is frequently and should be fast. 
-        float[] data2 = (float[]) data1;
+        int[] data2 = (int[]) data1;        // cc 2024-07 
         data1 = data2[ixArray[ixix]];
       } else if(clazz1 == Character.TYPE){
-        float[] data2 = (float[]) data1;
+        float[] data2 = (float[]) data1;    // TODO what is with the others?
         data1 = data2[ixArray[ixix]];
       } else if(clazz1 == Long.TYPE){
         float[] data2 = (float[]) data1;
